+++ src/cmd/cmd_common.rs
@@ -316,7 +316,7 @@ pub fn sel_pa_cmd(pa_sel: PaSel) -> [u8; 3] {
 pub fn set_rx_duty_cyle_cmd(rx_max_time: u32, cycle_time: u32, use_lora_cad: bool, dram_ret: u8) -> [u8; 9] {
     let mut cmd = [0u8; 9];
     cmd[0] = 0x02;
-    cmd[1] = 0x11;
+    cmd[1] = 0x10;
 
     cmd[2] |= ((rx_max_time >> 16) & 0xFF) as u8;
     cmd[3] |= ((rx_max_time >> 8) & 0xFF) as u8;